refactor: move cli update logic into runtime manager
