+++ crates/runtime-manager/Cargo.toml
@@ -16,6 +16,7 @@ chrono = { version = "0.4", default-features = false, features = ["clock"] }
 dirs = "5"
 reqwest = { version = "0.12", default-features = false, features = ["rustls-tls", "json"] }
 tokio = { workspace = true }
+tracing = "0.1"
 sysinfo = { version = "0.30", optional = true }
 
 [dev-dependencies]
